tests: create framework for automating QA tests
