+++ tests/smoke/qa_test.go
@@ -1,0 +1,6 @@
+package smoke
+
+import "testing"
+
+func testQA(t *testing.T) {
+}
